Keep MySql/Postgres prefixes on dialect-specific test class names
